+++ src/charm.py
@@ -30,7 +30,7 @@ from constants import (
 )
 from mysqlsh_helpers import MySQL
 from relations.db_router import DBRouterRelation
-from relations.provides import ProvidesRelation
+from relations.database import ProvidesRelation
 from relations.shared_db import SharedDBRelation
 from utils import generate_random_hash, generate_random_password
